backend/requirements.txt
kafka-python==2.0.2
requests==2.28.1
requests-toolbelt==0.9.1
